Restarting page project, lets create something more pleasing this time
